refactor(library): deepen library write path under library/
Move import, delete, songs metadata, and playlist ops into library
module; thin commands/import and playlist; remote mirror calls library
delete helpers (C6).
