+++ src-tauri/src/commands/import/mod.rs
@@ -1,37 +1,31 @@
-mod cover;
-pub(crate) mod delete;
-mod expand;
-mod ingest;
-mod preview;
-mod types;
+//! IPC adapter for library import / song write commands.
+//!
+//! Domain write path lives in `crate::library`. This module only binds
+//! Tauri state, opens the DB, and wraps remote mutation hooks.
 
-pub use cover::extract_embedded_cover_art_from_connection;
-pub use types::{
-    DeleteSongsFailure, DeleteSongsResult, ExpandedImportPaths, ExtractEmbeddedCoverArtFailure,
-    ExtractEmbeddedCoverArtResult, ImportCandidateDetails, ImportSongsOptions, SongProperties,
+// Re-export domain types/functions so existing `commands::import::…` paths
+// (tests, smoke) keep working during the transition.
+pub use crate::library::import::{
+    collect_expandable_import_paths, extract_embedded_cover_art_from_connection,
+    get_library_from_connection, import_songs_from_paths, import_songs_from_paths_with_options,
+    inspect_import_candidate, DeleteSongsFailure, DeleteSongsResult, ExpandedImportPaths,
+    ExtractEmbeddedCoverArtFailure, ExtractEmbeddedCoverArtResult, ImportCandidateDetails,
+    ImportSongsOptions, SongProperties,
+};
+pub use crate::library::songs::{
+    delete_songs as delete_songs_in_library, set_songs_instrumental as set_songs_instrumental_in_connection,
+    set_songs_language as set_songs_language_in_connection,
+    update_song_metadata as update_song_metadata_in_connection,
 };
 
 use crate::{
-    audio::decode,
     cache,
-    commands::error::{
-        database_error, internal_error, state_lock_error, CommandError, CommandResult,
-    },
+    commands::error::{database_error, state_lock_error, CommandError, CommandResult},
     commands::remote_library,
-    library::{error::LibraryError, ImportFailure, ImportSongsResult, Song},
-    library_root::LibraryRoot,
-    media_g::{self, MEDIA_G_PAIRED, MEDIA_G_ZIP},
+    library::{error::LibraryError, ImportSongsResult, Song},
     AppState,
 };
-use rusqlite::Connection;
-use std::collections::HashSet;
 use tauri::{AppHandle, State};
-
-use delete::delete_song_from_library;
-use expand::{build_selected_cdg_lookup, classify_import_paths, collect_expandable_import_paths};
-use ingest::{build_and_store_media_g_zip, build_and_store_song, try_extract_embedded_lyrics};
-use preview::{display_audio_format, inspect_import_candidate};
-use rayon::prelude::*;
 
 #[cfg(target_os = "macos")]
 use std::ffi::{c_char, CStr, CString};
@@ -46,6 +40,7 @@ pub fn import_songs(
     let library = state.library_root()?;
     let connection = cache::open_database(&library.database_path()).map_err(database_error)?;
 
+    // Remote Pre-Mutation Refresh / Publish Song: run_imported_songs_mutation
     remote_library::run_imported_songs_mutation(&state, &app_handle, || {
         import_songs_from_paths_with_options(
             &connection,
@@ -176,22 +171,11 @@ pub fn delete_songs(
         playback.current_song_id().map(|value| value.to_owned())
     };
 
-    let mut deleted_song_ids = Vec::new();
-    let mut failed = Vec::new();
-
-    for song_id in song_ids {
-        match delete_song_from_library(&connection, &library, &song_id) {
-            Ok(()) => deleted_song_ids.push(song_id),
-            Err(error) => failed.push(DeleteSongsFailure {
-                song_id,
-                error: CommandError::from(LibraryError::Internal(error.to_string())),
-            }),
-        }
-    }
+    let result = delete_songs_in_library(&connection, &library, &song_ids);
 
     if current_song_id
         .as_deref()
-        .is_some_and(|song_id| deleted_song_ids.iter().any(|deleted| deleted == song_id))
+        .is_some_and(|song_id| result.deleted_song_ids.iter().any(|deleted| deleted == song_id))
     {
         {
             let mut playback = state
@@ -209,10 +193,7 @@ pub fn delete_songs(
         *cdg_state = None;
     }
 
-    Ok(DeleteSongsResult {
-        deleted_song_ids,
-        failed,
-    })
+    Ok(result)
 }
 
 #[tauri::command]
@@ -238,110 +219,6 @@ pub fn extract_embedded_cover_art(
     )
 }
 
-pub fn import_songs_from_paths(
-    connection: &Connection,
-    library: &LibraryRoot,
-    paths: &[String],
-) -> ImportSongsResult {
-    import_songs_from_paths_with_options(connection, library, paths, &ImportSongsOptions::default())
-}
-
-pub fn import_songs_from_paths_with_options(
-    connection: &Connection,
-    library: &LibraryRoot,
-    paths: &[String],
-    options: &ImportSongsOptions,
-) -> ImportSongsResult {
-    let mut imported = Vec::new();
-    let mut failed = Vec::new();
-    let classified = classify_import_paths(paths);
-    let selected_cdg_by_stem = build_selected_cdg_lookup(&classified.cdg_paths);
-
-    // Phase 1: Read metadata and copy files in parallel (I/O bound).
-    // Phase 2: Insert into database sequentially (SQLite single-writer).
-    let prepared: Vec<_> = classified
-        .audio_paths
-        .par_iter()
-        .map(|audio_path| {
-            let result = build_and_store_song(
-                audio_path,
-                library,
-                &selected_cdg_by_stem,
-                &options.explicit_cdg_by_audio_path,
-                &options.skip_cdg_for_audio_paths,
-                &mut HashSet::<std::path::PathBuf>::new(),
-            );
-            (audio_path, result)
-        })
-        .collect();
-
-    // Collect consumed CDG source paths from successful imports.
-    // Uses the original source paths (not library-relative) so the standalone
-    // CDG check below can match against classified.cdg_paths.
-    let mut consumed_cdg_paths = HashSet::new();
-
-    for (audio_path, result) in prepared {
-        match result {
-            Ok(build_result) => {
-                if let Some(cdg_source) = build_result.consumed_cdg_source {
-                    consumed_cdg_paths.insert(cdg_source);
-                }
-                let song = build_result.song;
-                match cache::upsert_song(connection, &song) {
-                    Ok(()) => {
-                        try_extract_embedded_lyrics(connection, &song, library);
-                        imported.push(song);
-                    }
-                    Err(error) => failed.push(ImportFailure {
-                        path: audio_path.display().to_string(),
-                        error: database_error(error.to_string()),
-                    }),
-                }
-            }
-            Err(error) => failed.push(ImportFailure {
-                path: audio_path.display().to_string(),
-                error: CommandError::from(LibraryError::MediaReadFailed(error.to_string())),
-            }),
-        }
-    }
-
-    for zip_path in &classified.zip_paths {
-        match build_and_store_media_g_zip(zip_path, library) {
-            Ok(song) => match cache::upsert_song(connection, &song) {
-                Ok(()) => imported.push(song),
-                Err(error) => failed.push(ImportFailure {
-                    path: zip_path.display().to_string(),
-                    error: database_error(error.to_string()),
-                }),
-            },
-            Err(error) => failed.push(ImportFailure {
-                path: zip_path.display().to_string(),
-                error: CommandError::from(LibraryError::MediaReadFailed(error.to_string())),
-            }),
-        }
-    }
-
-    for cdg_path in &classified.cdg_paths {
-        if consumed_cdg_paths.contains(cdg_path) {
-            continue;
-        }
-
-        failed.push(ImportFailure {
-            path: cdg_path.display().to_string(),
-            error: CommandError::from(LibraryError::Internal(format!(
-                "standalone .cdg file {} does not have a matching audio track",
-                cdg_path.display()
-            ))),
-        });
-    }
-
-    ImportSongsResult { imported, failed }
-}
-
-pub fn get_library_from_connection(connection: &Connection) -> rusqlite::Result<Vec<Song>> {
-    cache::list_songs(connection)
-}
-
 #[tauri::command]
 pub fn update_song_metadata(
     state: State<'_, AppState>,
@@ -354,12 +231,12 @@ pub fn update_song_metadata(
     let connection = cache::open_database(&library.database_path()).map_err(database_error)?;
 
     remote_library::run_song_database_mutation(&state, &app_handle, &hash, || {
-        cache::update_song_title_artist(&connection, &hash, title.as_deref(), artist.as_deref())
-            .map_err(|e| database_error(e.to_string()))?;
-
-        cache::get_song_by_hash(&connection, &hash)
-            .map_err(|e| database_error(e.to_string()))?
-            .ok_or_else(|| database_error(format!("song with hash {hash} not found")))
+        update_song_metadata_in_connection(
+            &connection,
+            &hash,
+            title.as_deref(),
+            artist.as_deref(),
+        )
     })
 }
 
@@ -378,31 +255,6 @@ pub fn set_songs_instrumental(
     })
 }
 
-pub fn set_songs_instrumental_in_connection(
-    connection: &Connection,
-    song_ids: &[String],
-    instrumental: bool,
-) -> CommandResult<Vec<Song>> {
-    let mut updated_songs = Vec::with_capacity(song_ids.len());
-
-    for song_id in song_ids {
-        let updated = cache::update_song_instrumental(connection, song_id, instrumental)
-            .map_err(|error| database_error(error.to_string()))?;
-        if updated == 0 {
-            return Err(database_error(format!(
-                "song with hash {song_id} not found"
-            )));
-        }
-
-        let song = cache::get_song_by_hash(connection, song_id)
-            .map_err(|error| database_error(error.to_string()))?
-            .ok_or_else(|| database_error(format!("song with hash {song_id} not found")))?;
-        updated_songs.push(song);
-    }
-
-    Ok(updated_songs)
-}
-
 #[tauri::command]
 pub fn set_songs_language(
     state: State<'_, AppState>,
@@ -418,31 +270,6 @@ pub fn set_songs_language(
     })
 }
 
-pub fn set_songs_language_in_connection(
-    connection: &Connection,
-    song_ids: &[String],
-    language: Option<&str>,
-) -> CommandResult<Vec<Song>> {
-    let mut updated_songs = Vec::with_capacity(song_ids.len());
-
-    for song_id in song_ids {
-        let updated = cache::update_song_language(connection, song_id, language)
-            .map_err(|error| database_error(error.to_string()))?;
-        if updated == 0 {
-            return Err(database_error(format!(
-                "song with hash {song_id} not found"
-            )));
-        }
-
-        let song = cache::get_song_by_hash(connection, song_id)
-            .map_err(|error| database_error(error.to_string()))?
-            .ok_or_else(|| database_error(format!("song with hash {song_id} not found")))?;
-        updated_songs.push(song);
-    }
-
-    Ok(updated_songs)
-}
-
 #[tauri::command]
 pub fn get_song_properties(
     state: State<'_, AppState>,
@@ -451,167 +278,19 @@ pub fn get_song_properties(
     let library = state.library_root()?;
     let connection = cache::open_database(&library.database_path()).map_err(database_error)?;
 
+    // Ensure remote working-copy files exist before probing (command-layer only).
     let song = cache::get_song_by_hash(&connection, &song_id)
         .map_err(|e| database_error(e.to_string()))?
         .ok_or_else(|| database_error(format!("song with hash {song_id} not found")))?;
 
-    let Some(song_path) = song.file_path.as_deref() else {
-        return Err(CommandError::from(LibraryError::Internal(format!(
-            "song {} does not have a local file path",
-            song_id
-        ))));
-    };
     if song.is_remote() {
-        remote_library::ensure_remote_file_cached(&state.shell.app_data_dir, song_path)?;
+        if let Some(song_path) = song.file_path.as_deref() {
+            remote_library::ensure_remote_file_cached(&state.shell.app_data_dir, song_path)?;
+        }
         if let Some(cdg_path) = song.cdg_path.as_deref() {
             remote_library::ensure_remote_file_cached(&state.shell.app_data_dir, cdg_path)?;
         }
     }
-    let file_path = library.resolve(song_path);
-    let ext = song
-        .original_ext
-        .as_deref()
-        .or_else(|| file_path.extension().and_then(|e| e.to_str()))
-        .unwrap_or("bin");
 
-    let (decoded, file_size, format) = if song.media_g_container.as_deref() == Some(MEDIA_G_ZIP) {
-        let asset = media_g::inspect_zip_for_media_g(&file_path).map_err(|error| {
-            CommandError::from(LibraryError::MediaReadFailed(error.to_string()))
-        })?;
-        let decoded = decode::decode_bytes(asset.audio_bytes, ext).map_err(|e| {
-            internal_error(format!("failed to decode audio for {}: {}", song_id, e))
-        })?;
-        let file_size = std::fs::metadata(&file_path)
-            .map_err(|e| {
-                CommandError::from(LibraryError::MediaReadFailed(format!(
-                    "failed to open Media+G ZIP at {}: {}",
-                    file_path.display(),
-                    e
-                )))
-            })?
-            .len();
-        (
-            decoded,
-            file_size,
-            format!("{}+G ZIP", display_audio_format(ext)),
-        )
-    } else {
-        let decoded = decode::decode_file(&file_path).map_err(|e| {
-            internal_error(format!(
-                "failed to decode audio for {}: {}",
-                file_path.display(),
-                e
-            ))
-        })?;
-        let file_size = std::fs::metadata(&file_path)
-            .map_err(|e| {
-                CommandError::from(LibraryError::MediaReadFailed(format!(
-                    "failed to open audio file at {}: {}",
-                    file_path.display(),
-                    e
-                )))
-            })?
-            .len();
-        let format = if song.media_g_container.as_deref() == Some(MEDIA_G_PAIRED) {
-            format!("{}+G", display_audio_format(ext))
-        } else {
-            display_audio_format(ext).to_owned()
-        };
-        (decoded, file_size, format)
-    };
-
-    let bit_rate = if song.duration_ms > 0 {
-        let duration_secs = song.duration_ms as f64 / 1000.0;
-        Some(((file_size as f64 * 8.0) / duration_secs / 1000.0).round() as u32)
-    } else {
-        None
-    };
-
-    Ok(SongProperties {
-        format,
-        sample_rate: Some(decoded.sample_rate),
-        channels: Some(decoded.channels as u16),
-        bit_rate,
-        file_size,
-        duration_ms: song.duration_ms,
-        hash: song.hash,
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::expand::collect_expandable_import_paths;
-    use std::fs;
-    use tempfile::TempDir;
-
-    fn write_file(path: &std::path::Path) {
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent).expect("failed to create test directory");
-        }
-        fs::write(path, b"test").expect("failed to write test file");
-    }
-
-    #[test]
-    fn expands_importable_files_recursively() {
-        let tempdir = TempDir::new().expect("failed to create tempdir");
-        let root = tempdir.path();
-
-        write_file(&root.join("one.mp3"));
-        write_file(&root.join("two.lrc"));
-        write_file(&root.join("three.cdg"));
-        write_file(&root.join("nested/four.zip"));
-        write_file(&root.join("nested/deeper/five.flac"));
-        write_file(&root.join("nested/deeper/too/deep/hidden.mp3"));
-
-        let result = collect_expandable_import_paths(&[root.display().to_string()]);
-
-        assert_eq!(result.song_count, 3);
-        assert!(result.paths.iter().any(|path| path.ends_with("one.mp3")));
-        assert!(result.paths.iter().any(|path| path.ends_with("two.lrc")));
-        assert!(result.paths.iter().any(|path| path.ends_with("three.cdg")));
-        assert!(result
-            .paths
-            .iter()
-            .any(|path| path.ends_with("nested/four.zip")));
-        assert!(result
-            .paths
-            .iter()
-            .any(|path| path.ends_with("nested/deeper/five.flac")));
-        assert!(!result
-            .paths
-            .iter()
-            .any(|path| path.ends_with("nested/deeper/too/deep/hidden.mp3")));
-    }
-
-    #[test]
-    fn caps_recursive_import_scanning_depth() {
-        let tempdir = TempDir::new().expect("failed to create tempdir");
-        let root = tempdir.path();
-
-        write_file(&root.join("level-0.mp3"));
-        write_file(&root.join("level-1/level-1.mp3"));
-        write_file(&root.join("level-1/level-2/level-2.mp3"));
-        write_file(&root.join("level-1/level-2/level-3/level-3.mp3"));
-        write_file(&root.join("level-1/level-2/level-3/level-4/level-5/level-5.mp3"));
-
-        let result = collect_expandable_import_paths(&[root.join("level-1").display().to_string()]);
-
-        assert_eq!(result.song_count, 3);
-        assert!(result
-            .paths
-            .iter()
-            .any(|path| path.ends_with("level-1/level-1.mp3")));
-        assert!(result
-            .paths
-            .iter()
-            .any(|path| path.ends_with("level-1/level-2/level-2.mp3")));
-        assert!(result
-            .paths
-            .iter()
-            .any(|path| path.ends_with("level-1/level-2/level-3/level-3.mp3")));
-        assert!(!result
-            .paths
-            .iter()
-            .any(|path| path.ends_with("level-1/level-2/level-3/level-4/level-5/level-5.mp3")));
-    }
+    crate::library::songs::get_song_properties(&connection, &library, &song_id)
 }
